Initial Commit
replicated basic template of pytla.

created HA9 base class that on init reads a yaml file to create a bunch of
hidden functions for basic operations. These will be used as building blocks for
user facing functions.

created an error class which we will populate with errors defined in the manual
later.

also included the manual here for reference since there didnt seem to be a
single static site for it.

created yaml file based on the operation commands section of the manual